Added 'firing bullet' feature to Ship demo
On space keyup new bullets will be added to the 'bullets' array, each
one with its own 'impulse' vector. This is just a demo, in the future
this may be done creating a bullet objecto that has its own acceleration
vector inside.

Every bullet is removed from the array and destroyed when it goes
outside of the screen.