Added 'Monitor' class (see desscription)
With the very simple Monitor class you can create an on-screen monitor
with multiple entitled outputs that will display a value.

With this class you can reduce the use of the borwser JS Console that,
in some cases, may consume a lot of resources but also is no very handy
to watch values.